+++ src/plotter/plotter.h
@@ -4,7 +4,7 @@
 
 struct inhalt {
 	FILE *save;
-	float *numbers;
+	double *numbers;
 	int columns;
 	int rows;
 };
@@ -17,8 +17,8 @@ void close(struct inhalt *data);
 int sizeofinhalt(void);
 struct inhalt *init_inhalt(void);
 void close_inhalt(struct inhalt *data);
-void ins_numbers(struct inhalt *data, float *numbers, int columns, int rows);
-void copy_numbers(struct inhalt *data, float *numbers, int columns, int rows);
+void ins_numbers(struct inhalt *data, double *numbers, int columns, int rows);
+void copy_numbers(struct inhalt *data, double *numbers, int columns, int rows);
 struct inhalt *init_plotter(char *name, char *headerstring);
-void cycle_plotter(struct inhalt *data, float *xnumber, float *ynumber);
+void cycle_plotter(struct inhalt *data, double *xnumber, double *ynumber);
 void close_plotter(struct inhalt *data);
